+++ web-wallet/CHANGELOG.md
@@ -19,10 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change Holdings component design [#1361](https://github.com/dusk-network/rusk/issues/1361)
 - Change `fiatCurrency`, `locale`, `tokenCurrency`, `token` to required properties in Balance component [#1323](https://github.com/dusk-network/rusk/issues/1323)
 - Change `package.json` fields to reflect repo change [#1367](https://github.com/dusk-network/rusk/issues/1367)
+- Change `walletStore.js` to receive gasPrice and gasLimit when `transfer` , `stake`, `unstake` and `withdrawRewards` are called [#1353](https://github.com/dusk-network/rusk/issues/1353)
 
 ### Removed
 
 - Remove `fiat` property from Balance component [#1323](https://github.com/dusk-network/rusk/issues/1323)
+- Remove `gasSettings` store update from `dashboard/+page.svelte.js` [#1353](https://github.com/dusk-network/rusk/issues/1353)
 
 ### Fixed
 - Fix Changelog to point to issues [#1368](https://github.com/dusk-network/rusk/issues/1368)